Pulled logic out of LocationsListPageController
